+++ android/khajit/app/src/main/java/com/project/khajit_app/global/User.kt
@@ -12,7 +12,7 @@ object User {
     var bio: String? = ""
     var location: String? = ""
     var phone_number: Int? = 0
-    var iban_number: Int? = 0
+    var iban_number: Long? = 0
     var citizenship_number: String? = ""
     var last_changed_password_date: String? = ""
 }
